requirements.txt
# Runtime
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
jinja2==3.1.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
httpx==0.28.1
pydantic-settings==2.6.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
respx==0.23.1
